move convert_units to charge_tools
