+++ pbj/mesh/charge_tools.py
@@ -977,3 +977,112 @@ def generate_msms_mesh_import_tinker_multipoles(solute):
         p12scale,
         p13scale,
     )
+
+
+def convert_units(units, magnitude="potential", temperature=298.15):
+    r"""Compute the scalar conversion factor from atomic electrostatic units to a target unit system.
+
+    The input unit string is normalized and matched against the supported aliases for
+    electrostatic potentials, derivatives, energies, and forces. The function then returns
+    the corresponding conversion factor and a human-readable label.
+
+    Args:
+        units (str/object): Target unit identifier. Supported aliases include `mv`, `mv_a`,
+            `mv_m`, `v`, `volt`, `volts`, `v_m`, `v_a`, `kt_e`, `kt_a`, `kt`,
+            `kj_mol_e`, `kj_mol`, `kj_mola`, `kj_mol_a`, `kcal_mol_e`, `kcal_mol`,
+            `kcal_mola`, `kcal_mol_a`, `e_eps0_angs`, `e_eps0_ang`, `atomic`, and `pn`.
+            The `pn` alias is only valid when `magnitude='force'`.
+        magnitude (str, optional): Physical quantity being converted. Must be one of
+            `potential`, `d_potential`, `energy`, or `force`. Defaults to `potential`.
+        temperature (float, optional): Absolute temperature in Kelvin used for thermal energy
+            conversions. Defaults to 298.15.
+
+    Returns:
+        tuple: A tuple containing the conversion factor and a descriptive unit label.
+
+    Raises:
+        ValueError: If `pn` is requested for a magnitude other than `force`, or if the
+            requested magnitude is not recognized.
+    """
+    units = str(units).strip().lower().replace("-", "_").replace(" ", "_")
+    units = units.replace("__", "_")
+    magnitude = str(magnitude).strip().lower()
+
+    qe = 1.60217663e-19
+    eps0 = 8.8541878128e-12
+    ang_to_m = 1e-10
+    kb = 1.380649e-23
+    kT = kb * temperature
+    Na = 6.02214076e23
+
+    to_V = qe / (eps0 * ang_to_m)
+
+    if units in ["mv_m", "mv_a", "mv"]:
+        factor_base, label_base = to_V * 1000, "mV"
+    elif units in ["v", "volt", "volts", "v_m", "v_a"]:
+        factor_base, label_base = to_V, "V"
+    elif units in ["kt_e", "kt_a", "kt"]:
+        factor_base, label_base = to_V / (kT / qe), "kT/e"
+    elif units in ["kj_mol_e", "kj_mol", "kj_mola", "kj_mol_a"]:
+        factor_base, label_base = to_V * (qe * Na / 1000), "kJ/mol"
+    elif units in ["kcal_mol_e", "kcal_mol", "kcal_mola", "kcal_mol_a"]:
+        factor_base, label_base = to_V * (qe * Na / (4.184 * 1000)), "kcal/mol"
+    elif units in ["e_eps0_angs", "e_eps0_ang", "atomic"]:
+        factor_base, label_base = 1.0, "e/(eps0*A)"
+    elif units in ["pn"]:
+        if magnitude != "force":
+            raise ValueError(
+                f"Unit 'pN' is only valid for magnitude='force', not '{magnitude}'."
+            )
+        factor_base, label_base = (qe**2 / (eps0 * ang_to_m**2)) / 1e-12, "pN"
+        return factor_base, label_base
+    else:
+        if magnitude in ["potential", "d_potential"]:
+            print(
+                f"Warning: Unit '{units}' not recognized for {magnitude}. Defaulting to mV."
+            )
+            factor_base, label_base = to_V * 1000, "mV"
+        else:
+            print(
+                f"Warning: Unit '{units}' not recognized for {magnitude}. Defaulting to kcal/mol."
+            )
+            factor_base, label_base = to_V * (qe * Na / (4.184 * 1000)), "kcal/mol"
+
+    if magnitude == "potential":
+        if "mol" in label_base:
+            label_base += "/e"
+        return factor_base, label_base
+
+    elif magnitude in ["d_potential"]:
+        if units in ["e_eps0_angs", "e_eps0_ang", "atomic"]:
+            return 1.0, "e/(eps0*A**2)"
+        elif units in ["v_m"]:
+            return to_V / ang_to_m, "V/m"
+        elif units in ["mv_m"]:
+            return (to_V * 1000) / ang_to_m, "mV/m"
+        return factor_base, f"{label_base}/A"
+
+    elif magnitude == "energy":
+        if "mol" in label_base:
+            return factor_base, label_base
+        elif label_base == "kT/e":
+            return factor_base, "kT"
+        elif label_base == "e/(eps0*A)":
+            return 1.0, "e**2/(eps0*A)"
+        else:
+            return factor_base, f"{label_base}*e"
+
+    elif magnitude == "force":
+        if "mol" in label_base:
+            return factor_base, f"{label_base}/A"
+        elif label_base == "kT/e":
+            return factor_base, "kT/A"
+        elif label_base == "e/(eps0*A)":
+            return 1.0, "e**2/(eps0*A**2)"
+        else:
+            return factor_base, f"{label_base}*e/A"
+
+    else:
+        raise ValueError(
+            f"Magnitude '{magnitude}' not recognized. Choose from: potential, d_potential, energy, force."
+        )
